Renamed scripts to mjs so they can be treated as ESM but keep the website as it as before
