+++ package.json
@@ -2,6 +2,9 @@
   "name": "tarkov.dev",
   "version": "0.0.0",
   "private": true,
+  "engines": {
+    "node": ">=20"
+  },
   "scripts": {
     "analyze": "source-map-explorer 'build/static/js/*.js'",
     "prestart": "npm run prebuild",
@@ -10,12 +13,12 @@
     "test": "react-scripts test --env=jsdom",
     "prettier": "prettier --write \"src/**/*.{js,jsx,ts,tsx,json,css,scss,md}\"",
     "eject": "react-scripts eject",
-    "prebuild": "node scripts/update-props.js && node --experimental-loader ./scripts/custom-loader.mjs scripts/cache-api-data.mjs && node scripts/get-version.js && node scripts/get-contributors.js && node scripts/build-sitemap.js && node scripts/generate-thumbnails.js",
-    "postbuild": "node scripts/build-redirects.js && node scripts/critical.js",
+    "prebuild": "node scripts/update-props.mjs && node scripts/cache-api-data.mjs && node scripts/get-version.mjs && node scripts/get-contributors.mjs && node scripts/build-sitemap.mjs && node scripts/generate-thumbnails.mjs",
+    "postbuild": "node scripts/build-redirects.mjs && node scripts/critical.mjs",
     "stage": "npx rimraf build && npm run build && npm run preview",
     "preview": "npx serve build -l 3001 -s",
-    "critical": "node scripts/critical.js",
-    "cache-api-data": "node --experimental-loader ./scripts/custom-loader.mjs scripts/cache-api-data.mjs"
+    "critical": "node scripts/critical.mjs",
+    "cache-api-data": "node scripts/cache-api-data.mjs"
   },
   "devDependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
