+++ scripts/build-redirects.mjs
@@ -15,10 +15,10 @@ import fetch from "cross-fetch";
         process.exit(1);
     }
     const __dirname = new URL(".", import.meta.url).pathname;
-    let indexTemplate = fs.readFileSync(path.join(__dirname, '..', 'workers-site', 'index-template.js'), 'utf8');
+    let indexTemplate = fs.readFileSync(path.join(__dirname, '..', 'workers-site', 'index-template.mjs'), 'utf8');
     indexTemplate = indexTemplate.replace('REDIRECTS_DATA', JSON.stringify(redirects, null, 4));
 
     console.time('Write new data');
-    fs.writeFileSync(path.join(__dirname, '..', 'workers-site', 'index.js'), indexTemplate);
+    fs.writeFileSync(path.join(__dirname, '..', 'workers-site', 'index.mjs'), indexTemplate);
     console.timeEnd('Write new data');
 })();
