+++ scripts/custom-loader.mjs
@@ -2,7 +2,7 @@ import { URL } from "url";
 import { readFile } from "fs/promises";
 
 /**
- * This function forces .js files to be loades as ES modules
+ * This function forces .mjs files to be loades as ES modules
  * so the default export is a string containing the CSS stylesheet.
  */
 export async function load(url, context, defaultLoad) {
@@ -11,19 +11,19 @@ export async function load(url, context, defaultLoad) {
     }
 
     const forceConvert = [
-        'do-fetch-items.js',
-        'do-fetch-barters.js',
-        'do-fetch-crafts.js',
-        'do-fetch-hideout.js',
-        'do-fetch-maps.js',
-        'do-fetch-meta.js',
-        'do-fetch-traders.js',
-        'do-fetch-quests.js',
-        'do-fetch-bosses.js',
-        'flea-market-fee.js',
-        'camelcase-to-dashes.js',
-        'graphql-request.js',
-        'api-query.js',
+        'do-fetch-items.mjs',
+        'do-fetch-barters.mjs',
+        'do-fetch-crafts.mjs',
+        'do-fetch-hideout.mjs',
+        'do-fetch-maps.mjs',
+        'do-fetch-meta.mjs',
+        'do-fetch-traders.mjs',
+        'do-fetch-quests.mjs',
+        'do-fetch-bosses.mjs',
+        'flea-market-fee.mjs',
+        'camelcase-to-dashes.mjs',
+        'graphql-request.mjs',
+        'api-query.mjs',
     ];
 
     for (const fileName of forceConvert) {
