+++ scripts/generate-thumbnails.mjs
