+++ scripts/generate_items_thumbs.mjs
