+++ scripts/get-version.mjs
