+++ scripts/test-redirects.mjs
