+++ scripts/update-props.mjs
