+++ src/components/price-graph/index.js
@@ -11,7 +11,7 @@ import { useTranslation } from 'react-i18next';
 import Slider from 'rc-slider';
 
 import formatPrice from '../../modules/format-price.js';
-import { useQuery } from '../../modules/graphql-request.js';
+import { useQuery } from '../../modules/graphql-request.mjs';
 // import { getRelativeTimeAndUnit } from '../../modules/format-duration.js';
 
 import './index.css';
