+++ src/features/bosses/do-fetch-bosses.mjs
@@ -1,4 +1,4 @@
-import APIQuery from '../../modules/api-query.js';
+import APIQuery from '../../modules/api-query.mjs';
 
 class BossesQuery extends APIQuery {
     constructor() {
