+++ src/features/maps/do-fetch-maps.mjs
@@ -1,4 +1,4 @@
-import APIQuery from '../../modules/api-query.js';
+import APIQuery from '../../modules/api-query.mjs';
 
 class MapsQuery extends APIQuery {
     constructor() {
