+++ src/features/meta/index.js
@@ -3,7 +3,7 @@ import { useSelector, useDispatch } from 'react-redux';
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
 import equal from 'fast-deep-equal';
 
-import doFetchMeta from './do-fetch-meta.js';
+import doFetchMeta from './do-fetch-meta.mjs';
 import { langCode } from '../../modules/lang-helpers.js';
 import { placeholderMeta } from '../../modules/placeholder-data.js';
