+++ src/features/quests/index.js
@@ -3,7 +3,7 @@ import { useSelector, useDispatch } from 'react-redux';
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
 import equal from 'fast-deep-equal';
 
-import doFetchQuests from './do-fetch-quests.js';
+import doFetchQuests from './do-fetch-quests.mjs';
 import { langCode } from '../../modules/lang-helpers.js';
 import { placeholderTasks } from '../../modules/placeholder-data.js';
