+++ src/features/traders/index.js
@@ -4,7 +4,7 @@ import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
 import equal from 'fast-deep-equal';
 
 import { langCode } from '../../modules/lang-helpers.js';
-import doFetchTraders from './do-fetch-traders.js';
+import doFetchTraders from './do-fetch-traders.mjs';
 
 import { placeholderTraders } from '../../modules/placeholder-data.js';
